+++ src/app/move-popup/flavor-objective.component.js
@@ -1,0 +1,9 @@
+import template from './flavor-objective.html';
+import './flavor-objective.scss';
+
+export const FlavorObjectiveComponent = {
+  bindings: {
+    objective: '<'
+  },
+  template
+};